Average Annual Variability in Yield over the first 30 projection years (a performance metrics function of class PM)

Average Annual Variability in Yield over the first 30 projection years (a performance metrics function of class PM)

AAVY

Arguments

MSE

An object of class MSE

Format

An object of class PM of length 1.

Value

a matrix with n Management procedures (MSE@nMP) and nsim (MSE@nsim) columns from MSE

Examples

loadABT()

#> ABT-MSE objects loaded

AAVY(MSE_example)

#>           [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,] 0.0245504 0.01896356 0.02576633 0.02610702 0.03526414 0.02976245
#> [2,] 4.1343358 0.58903132 1.39701057 0.89481965 5.21211195 0.99643961
#> [3,] 0.1941974 0.17591194 0.15445884 0.73345322 0.21692089 0.18188427
#> [4,] 0.0419452 0.04976594 0.04104232 0.04337508 0.04468184 0.04438322
#>            [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#> [1,] 0.03161418 0.02582242 0.03151233 0.02944484 0.02595314 0.02425824
#> [2,] 2.42855906 1.11020574 0.89915960 2.43947592 4.65769568 3.36599497
#> [3,] 0.25753805 0.09718545 0.09337216 0.20071308 0.20113830 0.12318440
#> [4,] 0.09824884 0.03746294 0.04359260 0.05007778 0.04313760 0.03911957
#>           [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#> [1,] 0.02662330 0.02562547 0.02543771 0.02626367 0.02498106 0.02641012
#> [2,] 4.98045447 1.70134492 1.66182877 1.29127065 2.10189981 1.95677154
#> [3,] 0.15003025 0.55485747 0.29653237 0.12146148 0.29114511 0.28134233
#> [4,] 0.04084358 0.07302165 0.09047470 0.04022306 0.03489380 0.06721667
#>           [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#> [1,] 0.02394252 0.02962651 0.03171457 0.02562740 0.02581486 0.03089609
#> [2,] 2.90474228 1.01001177 2.19568023 1.80931884 4.45172190 1.78152461
#> [3,] 0.24811699 0.18664505 0.50211744 0.20604898 0.20803294 0.10409354
#> [4,] 0.05336829 0.05055466 0.04632332 0.04160694 0.09745628 0.04509056
#>           [,25]      [,26]      [,27]      [,28]      [,29]      [,30]
#> [1,] 0.02562327 0.03482252 0.02639762 0.02835907 0.02981271 0.02113057
#> [2,] 4.82561908 1.04710330 1.60729371 1.64509113 0.84011357 3.02444605
#> [3,] 0.32245992 0.28122385 0.22062415 0.22980965 0.12606399 0.18213742
#> [4,] 0.03776301 0.05326260 0.04877129 0.04847270 0.04803272 0.02991115
#>           [,31]        [,32]     [,33]      [,34]      [,35]      [,36]
#> [1,] 0.03929325 3.200315e-02 0.0285741 0.03992773 0.03532366 0.03399758
#> [2,] 2.03960670 8.817230e-01 1.6887730 1.64304317 2.23395454 1.75125154
#> [3,] 0.23299242 1.606164e+12 0.2186300 0.14253133 0.27900689 0.16769535
#> [4,] 0.11694241 8.586861e-02 0.0646343 0.32232346 0.04143624 0.04868878
#>           [,37]      [,38]        [,39]      [,40]      [,41]      [,42]
#> [1,] 0.02559347 0.02684311 2.458205e-02 0.02802391 0.02603984 0.02883592
#> [2,] 1.69745260 2.36899286 2.755458e+00 1.58524073 2.16553732 2.42403808
#> [3,] 0.20629764 0.32776715 2.637971e+13 0.19388442 0.29111472 0.16672946
#> [4,] 0.09305928 0.50829725 3.721793e-02 0.04745691 0.04655553 0.04545687
#>           [,43]      [,44]      [,45]      [,46]      [,47]      [,48]
#> [1,] 0.02933109 0.02677495 0.02781339 0.01921318 0.02911666 0.02606705
#> [2,] 3.87843925 2.28856579 2.49331735 1.11283437 3.03117866 1.89515420
#> [3,] 0.21809662 0.17957578 0.20347574 0.32065875 0.26103634 0.28640215
#> [4,] 0.04087376 0.06184264 0.04514945 0.35447403 0.04193007 0.04304656