An adaptive surplus production management procedure using INDEX 2 that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).

An adaptive surplus production management procedure using INDEX 2 that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).

Fadapt_i2

Arguments

x

a simulation number.

dset

a list of simulated data for use by management procedures.

yrsmth

the number of years over which to quantify slope in surplus production with biomass

gg

the responsiveness of the MP with respect to the slope

FMSY_M

the assumed ratio of FMSY to M

Format

An object of class MP of length 1.

Value

a TAC recommendation arising from x, dset.

Examples

Fadapt_i2(1,dset_example_East)

#> Estimate 
#>  5558314 

sapply(1:10,Fadapt_i2,dset_example_East)

#>  Estimate  Estimate  Estimate  Estimate  Estimate  Estimate  Estimate  Estimate 
#> 5558314.2 3852447.6 3759595.3 2911145.5 2152860.1 7669399.4 1867825.1 8688148.7 
#>  Estimate  Estimate 
#> 4092234.5  778672.3