An S4 class object that contains all MSE outputs

Name

The name of the MSE

Date

The date and time the object was created

Author

Who made the object

Notes

Anything noteworthy

PrimarySource

The primary origins of these data

nsim

The number of simulations

npop

number of stocks (e.g. 2, East-West)

nages

number of ages (e.g. 18: 1-17 and 18+)

nyears

number of years of simulation by statistical catch at length

nHyears

number of historical years for stock reduction analysis prior ot statistical catch at length

nsubyears

number of subyears (e.g. 4 quarters)

nareas

number of areas

proyears

number of projection years for closed-loop MSE analysis

nlen

number of length classes

lenbins

vector of length bins that define length classes

interval

how often (number of years) assessments/MPs are implemented (e.g. interval=3 means 2015 assessment, 2016-2017 no assessment, 2018 assessment etc)

nma

number of movement age classes

ma

an array of movement age classes [npop x nage]

mulen

a vector of mean length in each length bin [nage]

Cimp

a vector of imprecision in catch observations (lognormal sd) by simulation [nsim]

Cb

a vector representing the bias in catch observations

Cerr

a matrix of the catch mulitpliers applied to true simulated catch the product of Cimp and Cb [nsim x nyears]

Iimp

currently unused

Ib

currently unused

Ierr

currently unused

nCAAobs

a vector of the number of annual catch-at-age observations by simulation [nsim]

nCALobs

a vector of the number of annual catch-at-length observations by simulation [nsim]

Lcv

currently unused

Mb

a vector of biases in natural mortality rate by simulation [nsim]

Kb

a vector of biases in von-Bert growth parameter K by simulation [nsim]

t0b

a vector of biases in von-Bert growth parameter t0 by simulation [nsim]

Mb

a vector of biases in von-Bert maximum length parameter by simulation [nsim]

LFCb

currently unused

LFSb

currently unused

FMSYb

a vector of biases in fishing mortality rate at MSY by simulation [nsim]

FMSY_Mb

a vector of biases in the ratio of fishing mortality rate at MSY to natural mortality rate by simulation [nsim]

BMSY_B0b

currently unused

ageMb

a vector of biases in the age at 50 per cent maturity [nsim]

Dimp

currently unused

Db

currently unused

Derr

currently unused

Btimp

a vector of imprecision in current biomass observations (lognormal sd) by simulation [nsim]

Btb

a vector representing the bias in current biomass observations

Bterr

a matrix of the mulitpliers applied to true simulated biomass the product of Btimp and Btb [nsim x nyears]

Ftimp

currently unused

Ftb

currently unused

Fterr

currently unused

hb

currently unused

Reccbcv

currently unused

IMSYb

a vector of biases in the index at MSY by simulation [nsim]

IMSYb

a vector of biases in MSY (a reference catch level) by simulation [nsim]

IMSYb

a vector of biases in BMSY (a reference biomass level) by simulation [nsim]

C

a 4D array containing true simulated catches [MP x nsim x nstocks x nyears]

D

a 4D array containing true simulated stock depletion (SSB/SSB0) [MP x nsim x nstocks x nyears]

B_BMSY

a 4D array containing true simulated biomass relative to BMSY [MP x nsim x nstocks x nyears]

F_FMSY

a 4D array containing true simulated fishing mortality rate relative to FMSY [MP x nsim x nstocks x nyears]

B

currently unused

SSB

a 4D array containing true simulated spawning biomass [MP x nsim x nstocks x nyears]

SSB0

a 2D array containing unfished spawnign biomass by simulation and stock [nsim x nstocks]

SSB0proj

a 3D array containing unfished spawning biomass by simulation, stock and projection year [nsim x nstocks x proyears]

TAC

a 4D array containing the TAC recommendations [nsim x MP x stock x proyear]

nMPs

an integer number representing the number of MPs in the MSE

Snames

a character vector naming the stocks [nstocks]

area_defs

a list of area definitions for graphing (lons and lats describing the polygon)

areanams

a character vector of area names

MPs

a list object containing the names of the MPs [nMPs]