An S4 class object that contains all Operating Model Inputs in the format for M3 operating model estimation

Name

The name of the operating model input

Date

The date and time the object was created

Author

Who made the object

Notes

Anything noteworthy

PrimarySource

The primary origins of these data

nsim

The number of simulations

npop

number of stocks (e.g. 2, East-West)

nages

number of ages (e.g. 18: 1-17 and 18+)

nyears

number of years of simulation by statistical catch at length

nHyears

number of historical years for stock reduction analysis prior ot statistical catch at length

nsubyears

number of subyears (e.g. 4 quarters)

nareas

number of areas

proyears

number of projection years for closed-loop MSE analysis

nlen

number of length classes

lenbins

vector of length bins that define length classes

interval

how often (number of years) assessments/MPs are implemented (e.g. interval=3 means 2015 assessment, 2016-2017 no assessment, 2018 assessment etc)

nma

number of movement age classes

ma

an array of movement age classes [npop x nage]

Rectype

a matrix specifying the recruitment type (a stock can have multiple types of recruitment over the projected time period) by stock [type, stock]

h

a vector nstocks long specifying the steepness of the B-H stock recruitment curve (currently unused)

recgrad

a matrix specifying a possible future gradient in recruitment expressed as a per cent per year [stock x 2 (upper / lower)]

Recind

an array specifyign which historical recruitment type to use in each projected year [nsim x nstock x projection year]

Recpars

an array of stock recruitment parameters by recruitment type [nsim x nstock x rectype (Hockey stock / B-H) x 2 (h or jointpoint/R0)]

Reccv

a vector nstock long of recruitment variability by stock

AC

a vector nstock long of recruitment autocorrelation

Recsubyr

a vector nstock long specifying which subyear recruitment occurs in

Linfmu

a vector nstock long specifying the mean maximum length of individuals

Kmu

a vector nstock long specifying the mean growth parameter K

t0

a vector nstock long specifying the theoretical age at length zero

a

a vector nstock long, the weight-length parameter a W=aL^b

b

a vector nstock long, the weight-length parameter b W=aL^b

Len_age

an array of length at age by stock [stock x age]

Wt_age

an array of weight at age by stock [stock x age]

ageM

an array of maturity at age by stock [stock x age]

D

current stock depletion by stock (currently unused)

Size_area

a vector nareas long, the relative size of each area

mov

a very large array storing simulated movement by stock [sim x stock x age x subyear x from area x to area]

nfleets

the number of fishing fleets

iALK

an array representing the inverse age-length key (conditional probability of length given age) [sim x stock x year x age x length class]

hZ

an array of historical mortality rate estimates [stock, year, subyear, age, area]

HCobs

an array of historical catch observations [year, subyear, age, area]

Cobs

an array of catch observations [year, subyear, area, fleet]

E

an array of standardized effort (partial effort)[sim, fleet, year, subyear, area]

qE

a vector nfleets long, the catchability of the fleets

qI

a vector nI long of catchabilities by fishery independent index

qCPUE

a vector nCPUE long of catchabilities by fishery CPUE index

sel

an array of selectivities [sim, fleets, length class]

selpars

array of selectivity parameters (currently unused)

mat

array of maturity [sim, stock, age, year]

M

array of natural mortality rate [sim, stock, age, year]

Recdevs

of historical recruitment deviations [sim, stock, year]

R0

vector nstock long of unfished recruitment (currently unused - recpars)

muR

a vector nstock long of mean recruitment

MSY

a matrix of MSY values per simulation and stock [sim x stock]

BMSY

a matrix of Biomass at MSY values per simulation and stock [sim x stock]

VBMSY

a matrix of vulnerable Biomass at MSY values per simulation and stock [sim x stock]

SSBMSY

a matrix of spawning biomass at MSY values per simulation and stock [sim x stock]

FMSY

a matrix of fishing exploitation rate at MSY values per simulation and stock [sim x stock]

UMSY

a matrix of fishing harvest rate at MSY values per simulation and stock [sim x stock]

FMSYa

a matrix of apical (coresponding with age that is most vulnerable) fishing exploitation rate at MSY values per simulation and stock [sim x stock]

SSBMSY_SSB0

a matrix of spawning biomass at MSY relative to unfished, values per simulation and stock [sim x stock]

nydist

the number of years used to determine equilibrium spatial distribution

Snames

a vector nstock long naming each stock

area_defs

a list of area definitions for graphing (lons and lats describing the polygon)

areanams

a character vector of area names

seed

a random seed for generation of simulations to ensure reproducibility