An S4 class object that contains all Operating Model Inputs in the format for M3 operating model estimation

Name

The name of the operating model input

Date

The date and time the object was created

Author

Who made the object

Notes

Anything noteworthy

PrimarySource

The primary origins of these data

OMfactors

The various levels and factors of this operating model

years

The real year range of the operating model

Hyears

The historical year range of the operating model

areas

The codes of the various geographical areas of the model

areanams

Full names of the various geographical areas

area_defs

A list of the polygons representing each of the areas (for plotting)

Fleets

A list object containing the definitions of the fleets

L1

Richards growth curve parameter

L2

Richards growth curve parameter

K

Richards growth curve parameter

p

Richards growth curve parameter

ageM

Age at 50 percent maturity (not currently specified)

ageMsd

slope in the maturity ogive (not currently specified)

mat

maturity at age ogive [stock, age class, year]

lenbins

vector of length bins that define length classes

nHy

number of historical years (the period of stock reduction analysis)

ny

number of years for the statistical catch at length analysis

ns

number of subyears (e.g. 4 quarters)

np

number of stocks (e.g. 2, East-West)

na

number of ages (e.g. 18: 1-17 and 18+)

nr

number of areas

nf

number of fleets

nl

number of length classes

nRPT

number of recapture time periods (currently unused)

RPTind

the correct release subyear for a recapture in a given subyear at a particular recapture period (currently unused)

sdur

the relative size of each subyear duration e.g. c(0.25,0.25,0.25,0.25) for a quarterly model

nydist

number of years used to initialize the spatial distribution of the movement age classes

mulen

mean length of individuals in each length class

RDblock

a vector nyears long indicating which estimated recruitment to use in each year

nRD

the number of estimated recruitments allocated to year by RDblock

iALK

inverse age-length key, the conditional probability of a length class given an age class [stock, year, age class, length class]

lwa

length-weight parameter a W=aL^b(currently unused)

lwb

length-weight parameter b W=aL^b(currently unused)

len_age

an array of length-at-age [stock, age, year]

wt_age

an array of weight-at-age [stock, age, year]

Fec

an array of spawning biomass-at-age [stock, age, year]

spawns

the spawning season for each stock (a vector nstocks long)

canspawn

an array [nstock, nareas] indicating which areas each stock can spawn in (then calculated by the fraction of SSB in each area during spawns)

Ma

mortality at age [nstock, nages]

nCobs

the number of catch observations in the statistical catch at length phase

Cobs

a dataframe of catch observations, one row per observation [year, subyear, area, fleet observation]

nCPUEq

the number of CPUE indices used in the model fitting

nCPUEobs

the number of CPUE index observations used in model fitting

CPUEobs

a data frame of CPUE catch rate observations [year, stock, area, subyear, fleet, index] fitted against vulnerable biomass

nE

the number of partial F series (typically nfleets)

nEobs

the number effort observations (partial F's) one for each catch observations

Eobs

a data frame of effort observations one line per observation [year, subyear, area, fleet, effort]

nCLobs

the number of length observations

CLobs

a data frame of length observations [year, subyear]

HCobs

an array of historical catch observations [year x subyear x area x age]

RAI

an array of relative abundance - this is pass through data - it is the master index used to derive the standardized effort (partial F data Eobs)

nI

the number of fishery independent indices (e.g. 2 one spawning biomass survey in the GOM and MED)

nIobs

the number of fishery independent indices

Iobs

a data frame of fishery independent relative abundance observations [year, subyear, area, stock, index number, type (biomass/ssb), index]

nPSAT

the number of electronic tags of known stock of origin

PSAT

a data frame of electronic tag movements [stock, age, subyear, duration til recapture (subyears), from area, to area, number of tags]

nPSAT2

the number of electronic tags of unknown stock of origin

PSAT2

a data frame of electronic tag movements [age, subyear, duration til recapture (subyear), from area, to area, prob(stock1), prob(stock2)]

nTag

the number of conventional tagging movement observations

Tag

a data frame of conventional tagging observations (currently unused)

nSOOobs

the number of stock-of-origin observations

SOOobs

a data frame of stock-of-origin observations [stock, age, year, subyear, area, number of observations]

nsel

the number of selectivity types (not necessarily nf due to selectivity mirroring)

seltype

a vector nsel long specifying the type of selectivity (2 = logistic, 3 = Thompson (potentially domed))

selind

a vector nf long specifying the correct selectivity for each fleet

ratiolim

a vector indicating the ratio of precision to inflection point for the logistic selectivities

infleclim

the min/max for the position of selectivity inflection point relative to maximum length for logistic selectivities

nma

the number of movement age groups

ma

a vector na long specfying the correct movement for each age class

nMP

the number of movement parameters that are estimated

nmovind

the number of indices for assigning estimated movement parameters to the movement matrices

movind

a data frame containing the indexing (correct destination) of each movement parameter [stock, subyear, from area, to area]

nmov1

the number of indices for assiging a zero (not an estimated parameter, rows sum to 1 so the first movement is allocated a zero rather than an estimated parameter)

mov1

a data frame containing the indexing (correct destination) of a zero value in the movement matrices

movtype

the type of movement model (1: Gravity model, 2: Markov model)

CobsCV

a vector nf long, the catch observation error (lognormal sd)

CPUEobsCV

a vector nCPUEq long, the index observation error (lognormal sd)

IobsCV

a vector nI long, the fishery independent index observation error (lognormal sd)

RDCV

the recruitment deviation penalty (sigma R)(lognormal sd)

SSBprior

a vector np long, an optional prior on current spawning biomass

SSBpriorCV

the precision of the SSBprior (lognormal sd)

nLHw

the number of likelihood components (for weighting)

LHw

a vector nLHw long specifying the relative weight of the various data types 1 catch, 2 cpue, 3 FIindex, 4 Lcomp, 5 SOO, 6 PSAT, 7 PSAT2, 8 RecDev, 9 mov, 10 sel, 11 SRA, 12 SSB

muR_ini

a vector np long specifying an initial value of mean absolute recruitment

sel_ini

a matrix of initial values for selectivity [nf x nl]

selpar_ini

a matrix of initial value of selectivity parameters [nf x 3]

lnF_ini

a vector nCobs long specifying initial value for fishing mortality rates

ilnRD_ini

a vector of initial (year 1 disparities in age comp) recruitment deviations (currently unused)

lnRD_ini

a vector ny long of recruitment recruitment deviations

mov_ini

an array of initial movement [stock x age x subyear x area from x area to]

qCPUE_ini

a vector nCPUEq long of initial catchability estimates for CPUE indices

qI_ini

a vector nI long specifying initial value for catchabilities of fishery independent indices

D_ini

a vector np long specifying initial (comparative) depletion (SSB current / SSB unfished) estimates

ComplexRD

an integer specifying whether recruitments can be blocked or estiamted for each year (currently unused)

ComplexF

an integer specifying whether an F should be estimated for each catch observation rather than F=qE (currently unused)

nF

the number estimated F parameters (usually =nCobs)

nMPind

the number of fishery indices used for informing management procedures in subsequent MSE analysis (currently unused)

MPind

a data frame of fishery indices used for informing management procedures in subsequent MSE analysis (currently unused)

debug

a logical (0 or 1) value determing whether the model should be run for a single iteration to check for errors

verbose

a logical (0 or 1) value determing whether more complete information should be provided in each iteration of the model estimation

datacheck

a unique number for checking that data were read into the model correctly

CPUEnames

a character string nCPUEq long, recording the names of CPUE indices