An S4 class object that contains all observation model

Name

The name of the observation model object

Ccv

a 2 item vector specifying the range in catch observation error as lognormal sd (upper lower)

Cbcv

a 2 item vector specifying the maximum extend of catch bias as a lognormal sd

nCAAobs

a 2 item vector specifying the range in the number of annual catch-at-age observations

nCALobs

a 2 item vector specifying the range in the number of annual catch-at-length observations

Lcv

the variance in lengths by age class log normal sd (currently unused)

Ibeta

a 2 item vector representing the range of hyperstability in indices (currently unused)

Icv

a 2 item vector representing bounds the degree of uncertainty (noise) in relative abundance indices as a lognormal sd (upper lower)(currently unused)

Mbcv

a 2 item vector specifying the maximum extend of bias in natural mortality rate as a lognormal sd

Kbcv

a 2 item vector specifying the maximum extend of bias in growth rate K as a lognormal sd

t0bcv

a 2 item vector specifying the maximum extend of bias in growth parameter t0 as a normal cv

Linfbcv

a 2 item vector specifying the maximum extend of bias in maximum size Linf as a lognormal sd

LFCbcv

a 2 item vector specifying the maximum extend of bias in the length at first capture as a lognormal sd

LFSbcv

a 2 item vector specifying the maximum extend of bias in the length at full selection as a lognormal sd

FMSYbcv

a 2 item vector specifying the maximum extend of bias in simulated FMSY as a lognormal sd

FMSY_Mbcv

a 2 item vector specifying the maximum extend of bias in simulated FMSY / natural mortality rate as a lognormal sd

BMSY_B0bcv

a 2 item vector specifying the maximum extend of bias in simulated ratio of BMSY relative to unfished as a lognormal sd

ageMbcv

a 2 item vector specifying the maximum extend of bias in simulated FMSY / natural mortality rate as a lognormal sd

Dbcv

a 2 item vector specifying the maximum extend of bias in simulated depletion (current SSB relative to unfished) as a lognormal sd

Dcv

a 2 item vector specifying the range in interannual variability in depletion (current SSB relative to unfished) lognormal sd

Btbcv

a 2 item vector specifying the maximum extend of bias in simulated current biomass as a lognormal sd

Btcv

a 2 item vector specifying the range in interannual variability in current biomass as a lognormal sd

Ftbcv

a 2 item vector specifying the maximum extend of bias in simulated current exploitation rate as a lognormal sd

Ftcv

a 2 item vector specifying the range in interannual variability in current exploitation rate as a lognormal sd

hbcv

a 2 item vector specifying the maximum extend of bias in simulated steepness as a lognormal sd

Recbcv

a 2 item vector specifying the maximum extend of bias in simulated recent recruitment strength rate as a lognormal sd

IMSYbcv

a 2 item vector specifying the maximum extend of bias in simulated target relative abundance index as a lognormal sd

MSYbcv

a 2 item vector specifying the maximum extend of bias in simulated MSY (a target catch rate) as a lognormal sd

BMSYbcv

a 2 item vector specifying the maximum extend of bias in simulated BMSY (a target biomass level) as a lognormal sd

MPind

a data frame containing the indices that may be used by MPs in closed loop simulation [year, index number, index name, index, CV, stock, type (1: vulnerable biomass, 2: ssb), source, applies to model areas]

MPind_stats

a list containing the statistical properties of model fits (currently unused)