A surplus production management procedure using INDEX 2 that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).

A surplus production management procedure using INDEX 2 that attempts to identify the slope of changes in surplus production and biomass to locate a gradient of zero (a management procedure of class MP).

SPslope_i2

Arguments

x

a simulation number.

dset

a list of simulated data for use by management procedures.

yrsmth

the number of years over which to quantify slope in surplus production with biomass

alp

the SP gradient limits that lead to TAC changes

bet

the degree of responsiveness of the MP

Format

An object of class MP of length 1.

Value

a TAC recommendation arising from x, dset.

Examples

SPslope_i2(1,dset_example_East)

#> [1] 47541522

sapply(1:10,SPslope_i2,dset_example_East)

#>                     1          1          1                                1 
#> 47541521.9  2817571.1  3424322.3  4878827.9   840972.9 33579696.4  1882644.5 
#>                     1            
#>  9203684.3  2255701.4 10438302.3