An example simulated data set for the Western stock

Cobs

A matrix of [sim x nyear]

Iobs

A 3D array of observed relative abundance indices dimensions [sim x index x year]

K

A vector nsim long of von Bertalanffy growth parameter K estimates

Linf

A vector nsim long of growth parameter Linf estimates (maximum length)

t0

A vector nsim long of growth parameter t0 estimates

M

A matrix of natural mortality rate at age [sim x age]

Bt

A vector nsim long of current biomass estimates

MSY

A vector nsim long of MSY estimates

BMSY

A vector nsim long of BMSY estimates

UMSY

A vector nsim long of UMSY estimates

a

A vector nsim long of Length-weight conversion parameter a W=aL^b

b

A vector nsim long of Length-weight conversion parameter b W=aL^b

nages

An integer value specifying the total number of age classes

ageM

A vector nsim long of estimates of age at 50 per cent maturity

M

A matrix of maturity at age [sim x age]

Bt_PI

A vector nsim long of current biomass estimates known perfectly (PI: Perfect Information)

UMSY_PI

A vector nsim long of UMSY nown perfectly (PI: Perfect Information)

CAA

A 3D array of catch at age observations [sim x age x year]

CAL

A vector of lengths representing the start-end points of the length bins corresponding to CAL data]

MPrec

A vector nsim long of the previous recommendations of the management procedure