Calculate performance metrics.

Calculate performance metrics.

getperf(MSE)

Arguments

MSE

an object of class MSE

Value

a list n stocks long each with a data frame of performance metrics (columns) by management procedure (row) MSE.

Y10. Mean yield over the first 10 years of the projection

Y20. Mean yield over years 11-20 of the projection

Y30. Mean yield over years 21-30 of the projection

PGK. Probability of Green Kobe (F<FMSY AND B>BMSY) region over all years of the projection

POF. Probability of Over-Fishing (F>FMSY) over all years of the projection

POFed. Probability of Over-Fished status (B<BMSY) over all years of the projection

D10. Mean depletion (SSB relative to unfished) over the first 10 years of the projection

D20. Mean depletion (SSB relative to unfished) over years 11-20 of the projection

D30. Mean depletion (SSB relative to unfished) over years 21-30 of the projection

LD. Mean depletion (SSB relative to unfished) over all projected years

RSSB. Mean Spawning Stock Biomass in final projection year relative to zero fishing

LRSSB. Mean Spawning Stock Biomass over all projection years relative to zero fishing

AAVY. Mean Average Annual Variability in Yield over all projections

carat. weight of the diamond (0.2--5.01) ...

Examples

getperf(MSE_example)

#> $East
#>                      Y10          Y20          Y30      PGK       POF    POFed
#> ZeroC-ZeroC 3.949861e-07 3.519798e-07 3.050542e-07 82.29167  0.000000 17.70833
#> UMSY-UMSY   7.195752e+02 3.671963e+03 4.298851e+03 77.63889  3.263889 19.16667
#> DD_i4-DD_i2 4.400867e+03 7.775196e+03 9.082408e+03 12.43056 78.263889 64.16667
#> MeanC-MeanC 3.758140e+03 4.242738e+03 4.077607e+03 58.05556 36.597222 36.38889
#>                    D10       D20       D30          LD      RSSB     LRSSB
#> ZeroC-ZeroC 0.20180522 0.7805901 1.0302005 0.012858709 1.0000000 1.0000000
#> UMSY-UMSY   0.19278049 0.6584360 0.7579993 0.012801228 0.6989754 0.6780363
#> DD_i4-DD_i2 0.06156533 0.2215424 0.2146257 0.006099339 0.1700175 0.1248344
#> MeanC-MeanC 0.08837843 0.4129653 0.6038927 0.007259971 0.5648482 0.3049456
#>                     AAVY
#> ZeroC-ZeroC 2.803088e-02
#> UMSY-UMSY   2.226992e+00
#> DD_i4-DD_i2 5.830390e+11
#> MeanC-MeanC 7.436196e-02
#> 
#> $West
#>                      Y10          Y20          Y30       PGK      POF    POFed
#> ZeroC-ZeroC 1.077622e-07 7.593021e-08 6.961435e-08 88.125000  0.00000 11.87500
#> UMSY-UMSY   1.411174e+02 6.832478e+02 7.491973e+02 68.125000 15.20833 25.90278
#> DD_i4-DD_i2 8.092053e+02 6.039889e+02 3.911539e+02  3.541667 88.54167 86.80556
#> MeanC-MeanC 8.296087e+02 4.386892e+02 4.111078e+02 22.291667 62.91667 74.72222
#>                    D10        D20        D30          LD       RSSB     LRSSB
#> ZeroC-ZeroC 0.22268441 0.67663795 0.99787997 0.039432327 1.00000000 1.0000000
#> UMSY-UMSY   0.21077920 0.51325252 0.53735024 0.038511102 0.45467945 0.4509934
#> DD_i4-DD_i2 0.03741221 0.04601634 0.03008169 0.005492034 0.02076805 0.0179468
#> MeanC-MeanC 0.04034993 0.08720890 0.17727059 0.012480051 0.17242017 0.0808706
#>                     AAVY
#> ZeroC-ZeroC 8.476693e-02
#> UMSY-UMSY   1.346078e+00
#> DD_i4-DD_i2 3.248541e+10
#> MeanC-MeanC 1.546874e-01
#>